Rename __tests__ folder to test/cases and .spec to .test
