--- packages/form-for/src/__tests_tmp__/NestedChildren.spec.js
@@ -1,3 +1,0 @@
-describe("Nested children", () => {
-  it("empty", () => {});
-});
