+++ packages/form-for/src/test/cases/FieldAutoBind.test.js
@@ -1,12 +1,10 @@
 import React, { Component } from "react";
 import { mount } from "enzyme";
-import { Field, Form } from "../index";
-import Input from "../fixture/Input";
-import Select from "../fixture/Select";
+import { Field, Form } from "../../index";
+import "../fixture/registerComponents";
 
-describe("Field bind component", () => {
-  Field.bindComponent("text", Input);
-  Field.bindComponent("select", Select);
+describe("Custom schema", () => {
+  Field.enableAutoBind();
 
   const object = {
     schema: { name: { type: "text" }, role: { type: "select" } }
