+++ packages/form-for/src/test/cases/FieldDecorator.test.js
@@ -1,4 +1,4 @@
-import { field } from "../index";
+import { field } from "../../index";
 
 describe("Field decorator", () => {
   let testInstance;
