+++ packages/form-for/src/test/cases/FieldError.test.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { mount } from "enzyme";
-import { Field, Form } from "../index";
+import { Field, Form } from "../../index";
 import Input from "../fixture/Input";
 
 describe("Field error", () => {
